% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{discr.distance}
\alias{discr.distance}
\title{Distance}
\usage{
discr.distance(graphs, method = "F")
}
\arguments{
\item{graphs:}{[[n]][nroi, nroi] a collection of graphs for n subjects, where each graph is a square matrix.}

\item{method='F':}{the method for computing distances.
\describe{
   \item{'F'}{use the frobenius norm between pairs of graphs}
   \item{'H'}{use the hellinger distance between pairs of graphs}
}}
}
\value{
dist [n, n]: a matrix indicating the pairwise distances between all graphs passed in.
}
\description{
A function that returns a distance matrix given a collection of graphs. Note that this function uses the frobenius norm to compute the distance between all pairs of graphs.
}
\author{
Eric Bridgeford
}
