% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2discr.R
\name{discr.freq.obs2discr}
\alias{discr.freq.obs2discr}
\title{Frequency Discriminability}
\usage{
discr.freq.obs2discr(signal, ids, tr, lc = NaN, hc = NaN,
  spectrum = "amp", rank = FALSE, font.size = 15, method = "H")
}
\arguments{
\item{signal}{[[n]][nt, nroi] the signal for each of the n subjects, containing an array of nt observations for nroi rois.}

\item{ids}{[n] the ids for each scan corresponding to the signal from above.}

\item{tr=NaN}{[1] the repetition time of the dataset. NaN for none.}

\item{lc=NaN}{[1] the lower cutoff for highpass filtering. NaN for none.}

\item{hc=NaN}{[1] the high cutoff for low-pass filtering. NaN for none.}

\item{spec='amp'}{the spectrum to work with. 'amp' for amplitude, 'pow' for power.}

\item{rank=FALSE}{a boolean indicating whether to do unranked (FALSE) or ranked (TRUE) graphs, whereby all the edge-weights gare ranked before computing distances.}

\item{font.size=10}{the default font size for the plot text.}

\item{method='H'}{the method to use for the distance computation.
\describe{
  \item{'F'}{Frobenius norm between pairs of graphs}
  \item{'H'}{Hellinger distance between pairs of graphs}
}}
}
\value{
d [1] the discriminability statistic for the data.

dist [n, n] the distance matrix associated with the data.

distplot a plot of the distance matrix.

kdeplot a plot of the density estimate of the intra vs inter subject distances.

combinedplot a multiplot showing the dist plot and the kde plot.
}
\description{
A function that performs discriminability in the frequency domain.
}
\author{
Eric Bridgeford
}
