% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hell_dist.R
\name{discr.hell_dist}
\alias{discr.hell_dist}
\title{Hellinger Distances}
\usage{
discr.hell_dist(a, b)
}
\arguments{
\item{a}{the first matrix}

\item{b}{the second matrix}
}
\value{
h the hellinger distance
}
\description{
function that computes the hellinger distance between two matrices
}
\author{
Eric Bridgeford
}
\keyword{distance}
\keyword{hellinger}
