% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_edges.R
\name{discr.rank_edges}
\alias{discr.rank_edges}
\title{Rank Edges}
\usage{
discr.rank_edges(graphs, normalize = FALSE)
}
\arguments{
\item{graphs}{[[n]][nroi, nroi] the graphs to rank.}

\item{normalize=FALSE}{: a boolean indicating whether to normalize so the rank for a graph has values betweeen 0 and 1.}
}
\value{
rgraphs [[n]][nroi, nroi] the ranked graphs.
}
\description{
A function that ranks all the entries in each graph.
}
\author{
Eric Bridgeford
}
