% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_tests.R
\name{discr.tests.one_sample_test}
\alias{discr.tests.one_sample_test}
\title{Discriminability One Sample Permutation Test}
\usage{
discr.tests.one_sample_test(D, labels, nperm = 100, verbose = FALSE)
}
\arguments{
\item{D[nxn]:}{the distance matrix to run a permutation test for. An [nxn] matrix.}

\item{labels[n]:}{the labels organized appropriately with the distance matrix. Label 1 should correspond to the first column, 2 the second, and so on.}

\item{nrep=100:}{the number of permutations to perform.}

\item{verbose=FALSE:}{whether to print the itereation numbers.}
}
\value{
srel: the relative, unpermuted discriminability you want to see is significant.

null: the discriminability scores of the permuted data.

pval: the pvalue associated with the permutation test.
}
\description{
A function that permutes the labels of a distance matrix to obtain an empirical pvalue associated with whether the raw score is due to random chance.
}
\author{
Shangsi Wang
}
