% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2discr.R
\name{discr.time.obs2discr}
\alias{discr.time.obs2discr}
\title{Time Domain Discriminability}
\usage{
discr.time.obs2discr(signal, ids, rank = FALSE, graphs = FALSE,
  font.size = 15, method = "F")
}
\arguments{
\item{signal}{[[n]][nt, nroi]: the signal for each of the n subjects, containing an array of nt observations,  for nroi rois. Alternatively, can be the graphs for each of the subjects.}

\item{ids}{[n]: the ids for each scan corresponding to the signal from above.}

\item{rank=FALSE}{: a boolean indicating whether to do unranked (FALSE) or ranked (TRUE).}

\item{font.size=15:}{the default font size.}

\item{method='F'}{the method to use for the distance computation.
\describe{
  \item{'F'}{Frobenius norm between pairs of graphs}
  \item{'H'}{Hellinger distance between pairs of graphs}
}}
}
\value{
d [1]: the discriminability statistic for the data.

dist [n, n]: the distance matrix associated with the data.

distplot : a plot of the distance matrix.

kdeplot : a plot of the density estimate of the intra vs inter subject distances.

combinedplot : a multiplot showing the dist plot and the kde plot.
}
\description{
A function that performs discriminability in the time domain using the correlation.
}
\author{
Eric Bridgeford
}
