% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discr_perm_test.R, R/permute.R
\name{discr_perm_test}
\alias{discr_perm_test}
\alias{discr_perm_test}
\title{Discriminability Permutation Test}
\usage{
discr_perm_test(D, labels, nperm = 1000)

discr_perm_test(D, labels, nperm = 1000)
}
\arguments{
\item{D[nxn]:}{the distance matrix to run a permutation test for. An [nxn] matrix.}

\item{labels[n]:}{the labels organized appropriately with the distance matrix. Label 1 should correspond to the first column, 2 the second, and so on.}

\item{nperm=1000:}{the number of permutations to perform.}

\item{D[nxn]:}{the distance matrix to run a permutation test for. An [nxn] matrix.}

\item{labels[n]:}{the labels organized appropriately with the distance matrix. Label 1 should correspond to the first column, 2 the second, and so on.}

\item{nperm=1000:}{the number of permutations to perform.}
}
\value{
discrs: a list of the permutated discriminability scores.

d0: the score before any permutations are applied.

emp_discr: the empirical discriminability score.

pval: the pvalue associated with the permutation test.

scores: a list of the permutated discriminabiility scores.

s0: the score before any permutations are applied.

pval: the pvalue associated with the permutation test.
}
\description{
A function that permutes the labels of a distance matrix to obtain an empirical pvalue associated with whether the raw score is due to random chance.

A function that permutes the labels of a disctance matrix to obtain an empirical pvalue associated with whether the raw score is due
to random chance.
}
\author{
Eric Bridgeford

Eric Bridgeford
}
