% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance}
\usage{
distance(graphs)
}
\arguments{
\item{graphs}{[nroi, nroi, n]: a collection of graphs for n subjects, where each graph is a square matrix.}
}
\value{
dist [n, n]: a matrix indicating the pairwise distances between all graphs passed in.
}
\description{
A function that returns a distance matrix given a collection of graphs. Note that this function uses the frobenius norm to compute the distance between all pairs of graphs.
}
\author{
Gregory Kiar
}
