% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div2dist.R
\name{div2dist}
\alias{div2dist}
\title{Divergence to Distance
#'
a utility to convert divergence matrices to one matrix where each entry is the distance between each pair of matrices.}
\usage{
div2dist(div)
}
\arguments{
\item{div}{[[subs]][roi, rois]: the kullback-leibler divergence for each pair of rois.}

\item{dist}{[nsub, nsub]: the hellinger distance between each pair of subjects.}
}
\description{
Divergence to Distance
#'
a utility to convert divergence matrices to one matrix where each entry is the distance between each pair of matrices.
}
\author{
Eric Bridgeford
}
\keyword{distance}
\keyword{divergence,}
