% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq2div.R
\name{freq2div}
\alias{freq2div}
\title{Frequency to Divergence}
\usage{
freq2div(freq)
}
\arguments{
\item{freq}{[[n]][nbin, nroi] a list of n frequency spectrum signals with nbin bins and nroi rois.}
}
\value{
div_mtx [[n]][nroi, nroi] a list of n divergence matrices where each edge is a divergence between two pairs of rois.
}
\description{
a function that converts lists of frequency domain signals to divergence matrices.
}
\keyword{divergence}
\keyword{frequency,}
