% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_discr.R
\name{freq_discr}
\alias{freq_discr}
\title{Frequency Discriminability}
\usage{
freq_discr(signal, ids, tr, lc, spec = "amp", rank = FALSE, fsize = 15)
}
\arguments{
\item{signal}{[[n]][nt, nroi]: the signal for each of the n subjects, containing an array of nt observations for nroi rois.}

\item{ids}{[n]: the ids for each scan corresponding to the signal from above.}

\item{tr=2.5}{[1]: the repetition time of the dataset. NULL for none.}

\item{lc=0.01}{[1]: the lower cutoff for highpass filtering. NULL for none.}

\item{spec='amp'}{: the spectrum to work with. 'amp' for amplitude, 'pow' for power.}

\item{rank=FALSE}{: a boolean indicating whether to do unranked (FALSE) or ranked (TRUE).}

\item{fsize=10:}{the default font size for the plot text.}
}
\value{
d [1]: the discriminability statistic for the data.

dist [n, n]: the distance matrix associated with the data.

distplot : a plot of the distance matrix.

kdeplot : a plot of the density estimate of the intra vs inter subject distances.

combinedplot : a multiplot showing the dist plot and the kde plot.
}
\description{
A function that performs discriminability in the frequency domain.
}
\author{
Eric Bridgeford
}
