% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2freq.R
\name{highpass_fft}
\alias{highpass_fft}
\title{Highpass Fast Fourier Transform}
\usage{
highpass_fft(signal, tr = NaN, lc = NaN)
}
\arguments{
\item{signal}{[nbin, nroi]: an input signal.}

\item{tr}{[1]: the repetition time of the data (in seconds).}

\item{lc}{[1]: the lower cutoff (in Hz) for the fourier transform. fourier components below this will be set to 0.}
}
\value{
f [nbin/2, nroi]: the single-sided frequency domain of the signal per roi.
}
\description{
A utility function for computing the highpass-filtered, single-sided, real fourier transform.
}
\author{
Eric Bridgeford
}
