% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_div.R
\name{kl_div}
\alias{kl_div}
\title{Kullback Leibler Divergence}
\usage{
kl_div(a, b)
}
\arguments{
\item{a}{the first distribution}

\item{b}{the second distribution}
}
\description{
A function that computes the kullback leibler divergence for two distributions.
}
\author{
Eric Bridgeford
}
\keyword{divergence}
\keyword{kullback,}
\keyword{leibler,}
