% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2freq.R
\name{obs2amp}
\alias{obs2amp}
\title{Observations to Amplitude Spectrum}
\usage{
obs2amp(observations, tr = NaN, lc = 0.01, normalize = TRUE)
}
\arguments{
\item{observations}{[[n]][nt, nroi]: a list of signals for n subjects, each containing nt timesteps and nroi rois.}

\item{tr}{[1]: the repetition time (in sec) of the dataset. corresponds to the time to take a single timestep.}

\item{lc}{[1]: the lower cutoff (in Hz) below which the fourier domain components will be set to zero.}
}
\value{
amp_data [[n]][nt/2, nroi]: the amplitude spectrum of the dataset.
}
\description{
A utility to convert timeseries data to amplitude spectra.
}
\author{
Eric Bridgeford
}
