% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2corr.R
\name{obs2corr}
\alias{obs2corr}
\title{Observations to Correlation
a utility to convert timeseries to correlation matrices.}
\usage{
obs2corr(observations)
}
\arguments{
\item{observations}{[[subs]][timesteps, rois]: a list of observations for a particular subject.}
}
\value{
corr_data [[subs]][rois, rois]: the locally correlated roi timeseries.
}
\description{
Observations to Correlation
a utility to convert timeseries to correlation matrices.
}
\author{
Eric Bridgeford
}
