% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_timeseries.R
\name{open_timeseries}
\alias{open_timeseries}
\title{Open Timeseries
A utility to open timeseries information from a dataset.}
\usage{
open_timeseries(fnames, dataset_pos = 1, sub_pos = 2, run_pos = 3,
  exclude = FALSE)
}
\arguments{
\item{fnames}{[n]: a vector of filenames, with separation by underscores IE, dataset_subject_run_(other information).rds}

\item{run_pos}{[1]: the position of the run information in the filenames}

\item{dataset}{[1]: the 1-indexed position of the dataset id in the filenames, assuming - or _ delimiting in the filename}

\item{scan_pos}{[1]: the 1-indexed position of the subject id in the filenames, assuming - or _ delimiting in the filename}

\item{exclude=FALSE}{: a parameter idicating whether to ignore timeseries that contain missing data.}
}
\value{
ts [[subs]][nt, nroi]: the ts loaded from the specified file names. list of n subjects, each with nt timesteps and nroi rois

dataset [n]: a vector of the dataset ids for each subject.

subjects [n]: the subject ids

runs [n]: the run ids
}
\description{
Open Timeseries
A utility to open timeseries information from a dataset.
}
