% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_discr_perm_test.R
\name{paired_discr_perm_test}
\alias{paired_discr_perm_test}
\title{Paired Discriminability Permutation Test}
\usage{
paired_discr_perm_test(D1, D2, labels1, labels2, nperm = 1000,
  verbose = FALSE)
}
\arguments{
\item{D1[nxn]:}{the first distance matrix to run a permutation test for. An [nxn] matrix.}

\item{D2[nxn]:}{the second distance matrix to run a permutation test for. An [nxn] matrix.}

\item{labels1[n]:}{the first labels organized appropriately with the first distance matrix. Label 1 should correspond to the first column, 2 the second, and so on.}

\item{labels2[n]:}{the second labels organized appropriately with the second distance matrix. Label 1 should correspond to the first column, 2 the second, and so on.}

\item{nperm=1000:}{the number of permutations to perform.}
}
\value{
scores: a list of the permutated discriminability scores.

s0: the score before any permutations are applied.

emp_discr: the empirical discriminability score.

pval: the pvalue associated with the permutation test.
}
\description{
A function that permutes the rdf values associated with the classes in two distance matrices to determine whether a difference in discriminability score between the two matrices is significant.
}
\author{
Eric Bridgeford
}
