% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_square.R
\name{plot_square}
\alias{plot_square}
\title{Square plot}
\usage{
plot_square(mtx, title = "", xlabel = "ROI", ylabel = "ROI",
  legend = "metric", itype = "sq", fsize = 12)
}
\arguments{
\item{in}{[nroi, nroi] or [nt, nroi]: the input. Can either be a square matrix of featuresxfeatures, or an array of observationsxfeatures (see itype parameter).}

\item{title="":}{the title for the square plot.}

\item{xlabel="ROIs":}{the x label for the square plot.}

\item{ylabel="ROIs":}{the y label for the square plot.}

\item{legend="":}{the legend title for the square plot.}

\item{itype="sq":}{the shape of the input. If "sq", the plot will be generated as is for the input signal. If "ts", we will assume the input is observationsxfeatures, and will correlate the features first.}

\item{fsize=10:}{the default font size for the plot text.}
}
\value{
sqplot : a plot of the square.
}
\description{
A function that plots a square matrix. Good for things like Correlation matrices.
}
\author{
Eric Bridgeford
}
