% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Timeseries plot}
\usage{
plot_timeseries(ts, title = "", xlabel = "TRs", ylabel = "intensity",
  legend = "ROIs")
}
\arguments{
\item{signal}{[[n]][nt, nroi]: the signal for each of the n subjects, containing an array of nt observations for nroi rois.}

\item{title="":}{the title for the timeseries plot.}

\item{xlabel="TRs":}{the x label for the timeseries plot.}

\item{ylabel="intensity":}{the y label for the timeseries plot.}

\item{legend="ROIs":}{the legend title for the timeseries plot.}
}
\value{
tsplot : a plot of the timeseries.
}
\description{
A function that plots a timeseries.
}
\author{
Eric Bridgeford
}
