% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computerank.R
\name{rank_matrices}
\alias{rank_matrices}
\title{Rank Matrices}
\usage{
rank_matrices(graphs, normalize = FALSE)
}
\arguments{
\item{graphs}{[nroi, nroi, n]: the graphs to rank.}

\item{normalize=FALSE}{: a boolean indicating whether to normalize so the rank for a graph sums to 1.}
}
\value{
rg [nroi, nroi, n]: the ranked graphs.
}
\description{
A function that ranks all the entries in each graph.
}
\author{
Gregory Kiar
}
