% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{rdf}
\alias{rdf}
\title{Reliability Density Function}
\usage{
rdf(dist, ids)
}
\arguments{
\item{dist}{[n, n]: a distance matrix for n subjects.}

\item{ids}{[n]: a vector containing the subject ids for each subject.}
}
\value{
rdf [n]: the reliability per subject.
}
\description{
A function for computing the reliability density function of a dataset.
}
\author{
Eric Bridgeford and Gregory Kiar
}
