% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_square.R
\name{plot_square}
\alias{plot_square}
\title{Square plot}
\usage{
plot_square(mtx, title = "", xlabel = "TRs", ylabel = "intensity",
  legend = "ROIs")
}
\arguments{
\item{signal}{[[n]][nt, nroi]: the signal for each of the n subjects, containing an array of nt observations for nroi rois.}

\item{title="":}{the title for the square plot.}

\item{xlabel="ROIs":}{the x label for the square plot.}

\item{ylabel="ROIs":}{the y label for the square plot.}

\item{legend="":}{the legend title for the square plot.}
}
\value{
sqplot : a plot of the square.
}
\description{
A function that plots a square matrix. Good for things like Correlation matrices.
}
\author{
Eric Bridgeford
}
