% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_discr.R
\name{time_discr}
\alias{time_discr}
\title{Time Domain Discriminability}
\usage{
time_discr(signal, ids, rank = FALSE, graphs = FALSE, fsize = 15)
}
\arguments{
\item{signal}{[[n]][nt, nroi]: the signal for each of the n subjects, containing an array of nt observations,  for nroi rois. Alternatively, can be the graphs for each of the subjects.}

\item{ids}{[n]: the ids for each scan corresponding to the signal from above.}

\item{rank=FALSE}{: a boolean indicating whether to do unranked (FALSE) or ranked (TRUE).}

\item{graphs=FALSE:}{a boolean indicating whether the inputs are already graphs or not.}

\item{fsize=15:}{the default font size.}
}
\value{
d [1]: the discriminability statistic for the data.

dist [n, n]: the distance matrix associated with the data.

distplot : a plot of the distance matrix.

kdeplot : a plot of the density estimate of the intra vs inter subject distances.

combinedplot : a multiplot showing the dist plot and the kde plot.
}
\description{
A function that performs discriminability in the time domain using the correlation.
}
\author{
Eric Bridgeford
}
